requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
azure-ai-documentintelligence==1.0.2
    # via markitdown
azure-core==1.38.2
    # via
    #   azure-ai-documentintelligence
    #   azure-identity
azure-identity==1.25.2
    # via markitdown
beautifulsoup4==4.13.4
    # via
    #   markdownify
    #   markitdown
certifi==2025.6.15
    # via requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.2
    # via
    #   markitdown
    #   pdfminer-six
    #   requests
click==8.2.1
    # via magika
cobble==0.1.4
    # via mammoth
coloredlogs==15.0.1
    # via onnxruntime
cryptography==45.0.4
    # via
    #   azure-identity
    #   msal
    #   pdfminer-six
    #   pyjwt
darkdetect==0.8.0
    # via pyside6-fluent-widgets
defusedxml==0.7.1
    # via markitdown
et-xmlfile==2.0.0
    # via openpyxl
flatbuffers==25.2.10
    # via onnxruntime
humanfriendly==10.0
    # via coloredlogs
idna==3.10
    # via requests
isodate==0.7.2
    # via azure-ai-documentintelligence
lxml==6.0.0
    # via
    #   markitdown
    #   python-pptx
magika==0.6.2
    # via markitdown
mammoth==1.11.0
    # via markitdown
markdownify==1.1.0
    # via markitdown
markitdown==0.1.2
    # via markitdowngui (pyproject.toml)
mpmath==1.3.0
    # via sympy
msal==1.35.1
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==1.3.1
    # via azure-identity
numpy==2.2.6
    # via
    #   magika
    #   onnxruntime
    #   pandas
onnxruntime==1.22.0
    # via magika
openpyxl==3.1.5
    # via markitdown
packaging==25.0
    # via
    #   onnxruntime
    #   pytesseract
pandas==2.3.0
    # via markitdown
pdfminer-six==20250506
    # via markitdown
pillow==11.3.0
    # via
    #   pytesseract
    #   python-pptx
protobuf==6.31.1
    # via onnxruntime
pycparser==2.22
    # via cffi
pyjwt==2.11.0
    # via msal
pypdfium2==5.5.0
    # via markitdowngui (pyproject.toml)
pyside6==6.9.1
    # via
    #   markitdowngui (pyproject.toml)
    #   pyside6-fluent-widgets
pyside6-addons==6.9.1
    # via pyside6
pyside6-essentials==6.9.1
    # via
    #   pyside6
    #   pyside6-addons
pyside6-fluent-widgets==1.11.0
    # via markitdowngui (pyproject.toml)
pysidesix-frameless-window==0.7.7
    # via pyside6-fluent-widgets
pytesseract==0.3.13
    # via markitdowngui (pyproject.toml)
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.1
    # via magika
python-pptx==1.0.2
    # via markitdown
pytz==2025.2
    # via pandas
requests==2.32.4
    # via
    #   azure-core
    #   markitdown
    #   msal
shiboken6==6.9.1
    # via
    #   pyside6
    #   pyside6-addons
    #   pyside6-essentials
six==1.17.0
    # via
    #   markdownify
    #   python-dateutil
soupsieve==2.7
    # via beautifulsoup4
sympy==1.14.0
    # via onnxruntime
typing-extensions==4.14.0
    # via
    #   azure-ai-documentintelligence
    #   azure-core
    #   azure-identity
    #   beautifulsoup4
    #   python-pptx
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
xlrd==2.0.2
    # via markitdown
xlsxwriter==3.2.5
    # via python-pptx
